Start to move each solver to their separate dir
