+++ regression/stp.test.cpp
@@ -2,7 +2,7 @@
 #include "tests.h"
 
 #include <catch2/catch.hpp>
-#include <stpsolver.h>
+#include <stp/stpsolver.h>
 
 TEST_CASE("Simple STP test", "[STP]") {
   // Create STP Solver
