+++ src/CMakeLists.txt
@@ -1,4 +1,5 @@
 include_directories(${CMAKE_BINARY_DIR}/src)
+include_directories(${CMAKE_SOURCE_DIR}/src)
 
 #-----------------------------------------------------------------------------#
 # Macro to add solver files to compile
@@ -7,11 +8,13 @@ function(compile_solver)
     string(TOLOWER ${name} lower_name)
 
     if(CAMADA_ENABLE_SOLVER_${name})
-      if(NOT EXISTS "${CMAKE_SOURCE_DIR}/src/${lower_name}solver.cpp")
+      if(NOT EXISTS "${CMAKE_SOURCE_DIR}/src/${lower_name}/${lower_name}solver.cpp")
         message(FATAL_ERROR "Could not find backend implementation file ${CMAKE_SOURCE_DIR}/src/${lower_name}solver.cpp")
       endif()
 
-      list(APPEND solver_src_files ${lower_name}solver.cpp)
+      include_directories(${CMAKE_SOURCE_DIR}/src/${lower_name}/)
+
+      list(APPEND solver_src_files ${lower_name}/${lower_name}solver.cpp)
     endif()
   endforeach()
