+++ src/boolector/boolectorsolver.cpp
