+++ src/boolector/boolectorsolver.h
