+++ src/cvc4/cvc4solver.cpp
