+++ src/cvc4/cvc4solver.h
