+++ src/mathsat/mathsatsolver.cpp
