+++ src/mathsat/mathsatsolver.h
