+++ src/stp/stpsolver.cpp
