+++ src/stp/stpsolver.h
