+++ src/yices/yicessolver.cpp
