+++ src/yices/yicessolver.h
