+++ src/z3/z3solver.h
